feat(ui): allow agent annotations on the diff viewer (#213)
The Files diff viewer had no comment gutter, so you couldn't add agent
annotations on a diff the way you can in the plain editor. MergeView
builds each pane from `config.a/b.extensions`, so the gutter has to be
threaded through that shape (a pre-built EditorState would silently drop
it). Add optional `extraExtensions` + `onViewReady` to CmDiffEditor,
installed on the modified pane; extract the editor's comment
orchestration into a shared `useCommentGutter` hook; and add a
`CmDiffEditorWithComments` wrapper that DiffTab now renders.
